requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
black==23.1.0
build==0.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
Flask==2.2.3
frozendict==2.3.6
frozenlist==1.3.3
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
pip-tools==6.12.3
platformdirs==3.2.0
pycparser==2.21
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
requests-futures==1.0.0
six==1.16.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
yarl==1.8.2
yfinance==0.2.14
